==> student_submissions/projects/mcnulty/dan_hui/Notebooks/02_Clean_CollectionData.ipynb <==
# 02_Cleaning_Collection

Author: Daniel Hui

License: MIT

This cleans the columns of data and also removes unecessary columns

import pandas as pd
import numpy as np

#Global Variable
month = "jan"

#Load inventory 
inventory = pd.read_csv(F'../01_Data/02_Truncated/Library_Collection_Inventory_{month}_2018.csv',index_col=0)

inventory.head(3)

### Cleaning Functions

#This definition cleans all the junk in the publication year column
def clean_year(row):
    try:
        row = row.replace('[','')
        row = row.replace(']','')
        row = row.replace('?','')
        row = row.replace('c','')
        row = row.replace('.','')
        row = row.replace('-','')
        row = row.replace('©','')
        row = row.replace('between','')
        row = row.replace(' ','')
        row = row.replace(',','')
        row = row.replace('thatis','')
        row = row.replace('<','')
        row = row.replace('>','')
        row = row.replace('In','')
        row = row.replace('(','')
        row = row.replace(')','')
        if (row[0:2] in ['16','17','18','19','20']) & (len(row) >= 4) :      #if the first part looks 
            row = row[0:4]                                                   #like a year, grab it
        elif (row[-4:-2] in ['16','17','18','19','20']) & (len(row) >= 4) :  #if the last part looks 
            row = row[-4:0]                                                  #like a year,grab it
        else: row = np.nan                                                   #anything else is not a year
        
        try: 
            int(row)             #if what you've caught is a number, return it
            return row
        except: return np.nan    #otherwise return nan
    
    except: return np.nan

#This definition cleans all the junk in the ISBN column. There are often two ISBNs and only need one
def clean_isbn(row):
    try:
        return row.split(',')[0]
    except: return np.nan

### Data Cleaning

inventory['PublicationYear'] = inventory['PublicationYear'].apply(clean_year)
inventory['ISBN'] = inventory['ISBN'].apply(clean_isbn)

### Keep Needed Columns & Remove Duplicates

inventory = inventory[['BibNum','Title','ISBN','PublicationYear','Subjects','ItemType',
                       'ItemCollection','FloatingItem','ItemLocation',"ItemCount"]]
inventory = inventory.drop_duplicates()

### Export

inventory.to_csv(f'../01_Data/03_Cleaned/Library_Collection_Inventory_{month}_2018_clean.csv')
