==> student_submissions/projects/mcnulty/dan_hui/Notebooks/03_Merge_Branch_Libraries.ipynb <==
# 03_Merge_Branch_Libraries

Author: Daniel Hui

License: MIT

This Notebook merges a few sets of library branch information together

import pandas as pd
import numpy as np

### I. Branch Data

#### A. Load Book Inventory and Aggregate by Branch

#Load inventory 
inventory = pd.read_csv('../01_Data/03_Cleaned/Library_Collection_Inventory_jan_2018_clean.csv',index_col=0)

inventory_branch = inventory[~inventory["ISBN"].isnull()]                         #remove missing ISBNs 
inventory_branch = inventory_branch[["ItemLocation","ISBN"]].drop_duplicates()    #isol. only 2 cols & rem. dup
inventory_branch = inventory_branch.groupby(by="ItemLocation").count().sort_values(["ISBN"])
inventory_branch = inventory_branch.reset_index()
inventory_branch = inventory_branch.rename({"ItemLocation":"Code","ISBN":"Titles"},axis="columns")

#### B. Load Branch Addresses

#Load branch info
branch = pd.read_csv('../01_Data/02_Truncated/Library_Locations.csv',index_col=0)

combined_branch = branch.merge(inventory_branch,on='Code',how="left")

combined_branch.head()

#### C. Load Branch Latitude Longitude

#Load branch lat/long
branch_lat_long = pd.read_csv('../01_Data/01_Source/Library_Locations_Lat_Long.csv',index_col=0)
branch_lat_long = branch_lat_long.drop(["Description"],axis=1)

combined_branch = combined_branch.merge(branch_lat_long,on='Code',how="left")  #Merge! 
combined_branch.head()

#### D. Export to CSV

combined_branch.to_csv('../01_Data/03_Cleaned/Library_Branch_Merged.csv')
