==> student_submissions/projects/mcnulty/dan_hui/Notebooks/04_Split_CheckoutData_Feature_Target.ipynb <==
##  06_Split_CheckoutData_Feature_Target

Author: Daniel Hui

License: MIT

This notebook splits the checkout dataset into features and also checkout targets

import pandas as pd

### Load Checkout Data

#Load 2018 Checkouts
checkouts_18df = pd.read_csv('/Users/dhui/Downloads/01_Source_Data/Checkouts_By_Title_Data_Lens_2018.csv',index_col=0)

#Load 2017 Checkouts
checkouts_17df = pd.read_csv('/Users/dhui/Downloads/01_Source_Data/Checkouts_By_Title_Data_Lens_2017.csv',index_col=0)

checkouts_df = pd.concat([checkouts_17df, checkouts_18df])

checkouts_df.info()   #There are 11,125,504 checkout records from Jan 1 2017 to date

checkouts_df.describe()

checkouts_df.head(2)

### Limit Checkout Data to Circulating Books

#load book codes of ItemCollection codes for non-reference books
book_codes = pd.read_csv("../01_Data/03_Cleaned/ItemCollection_Book_Codes.csv",index_col=0,header=None)
book_codes = book_codes[1]

#truncate rows by keeping only items that are in the book code list
circulating_books_df = checkouts_df[checkouts_df["Collection"].isin(book_codes)]

circulating_books_df.shape   #7,106,862 checkouts 01/01/17 to Date for circulating books

### Keep Necessary Columns

circulating_books_df = circulating_books_df[["CheckoutYear","BibNumber","Collection","CheckoutDateTime"]]
circulating_books_df = circulating_books_df.reset_index(drop = True)

circulating_books_df.head(2)

### Clean DateTime to YearMonth

circulating_books_df["month"] = circulating_books_df["CheckoutDateTime"].apply(lambda x: x.split("/")[0])

circulating_books_df["YearMonth"] = circulating_books_df["CheckoutYear"].apply(lambda x: str(x)) + circulating_books_df["month"]

circulating_books_df.head(2)

circulating_books_df = circulating_books_df[["BibNumber","YearMonth","CheckoutDateTime"]]

circulating_books_df.head(2)

circulating_books_df.shape

### Export Target Set

#### 1st Half 2018 Target Set

Export books that were checked out in 2018 in the first half

date_list = ["201801", "201802", "201803", "201804", "201805", "201806"]
target_2018half_df = circulating_books_df[circulating_books_df["YearMonth"].isin(date_list)]
target_2018half_df = target_2018half_df["BibNumber"]
target_2018half_df = target_2018half_df.drop_duplicates()
target_2018half_df = target_2018half_df.reset_index(drop=True)

target_2018half_df.shape #215,873 unique titles checked out Q1-3 2018

target_2018half_df.to_csv("../01_Data/05_Target/Checkout_Set/18_Half.csv")

### Export Feature Set

Export books that were checked out in all of 2017 and leave all duplicate records in so we can later count hoe many times a book was checked out in different time increments

date_list = ["201701", "201702", "201703", "201704", "201705", "201706", 
             "201707", "201708", "201709", "201710", "201711", "201712"]
feature_2017_df = circulating_books_df[circulating_books_df["YearMonth"].isin(date_list)]

feature_2017_df.head()

feature_2017_df.shape 

feature_2017_df.to_csv("../01_Data/06_Features/2017_Checkouts.csv")
