==> student_submissions/projects/mcnulty/dan_hui/Notebooks/05_Create_Targets_CheckoutData.ipynb <==
## 07_Create_Targets_CheckoutData

Author: Daniel Hui

License: MIT

This notebook will create binary checkout targets by comparing books checked out with total books in the inventory.

import pandas as pd

### Import Library Inventory

inventory_df = pd.read_csv('../01_Data/03_Cleaned/Library_Collection_Inventory_jan_2018_clean.csv',index_col=0)

all_books = inventory_df["BibNum"].to_frame()
all_books = all_books.drop_duplicates()
all_books = all_books.reset_index(drop=True)

all_books.head(5)

all_books.shape   # There are 397,147 unique titles in the Seattle Public Library Collection

### Export: 1st Half 2018 Set

checkouts = pd.read_csv('../01_Data/05_Target/Checkout_Set/18_Half.csv',index_col=0,header=None)
checkouts = checkouts[1]
checkouts = checkouts.reset_index(drop=True)
checkouts = checkouts.to_frame()
checkouts = checkouts.rename({1:"BibNum"},axis=1)  # Rename Column
checkouts["Checkout"] = 0                          # Fill in dummy variable of 1 for being checked out in Q1
checkouts.shape                                    # 215,873 unique titles were checked out in 1st half of 2018

Half_target_df = all_books.merge(checkouts, on="BibNum",how='left')   #Merge to attach checked out 1
Half_target_df.head(4)

Half_target_df.shape

Half_target_df.describe()

Half_target_df = Half_target_df.fillna(1)
Half_target_df.head(4)

Half_target_df.describe()

Half_target_df.to_csv('../01_Data/05_Target/18_Half_Target.csv')
