==> student_submissions/projects/mcnulty/dan_hui/Notebooks/08_Create_Features-Location_CollectionData.ipynb <==
## 08_Create_Features-Location_CollectionData

Author: Daniel Hui

License: MIT

This notebook creates features from the collection dataset

import pandas as pd
pd.set_option('display.max_columns', 500)

#### Load Inventory and only keep relevant Columns

inventory_df = pd.read_csv('../01_Data/03_Cleaned/Library_Collection_Inventory_jan_2018_clean.csv',index_col=0)

inventory_df.head()

location_df = inventory_df[["BibNum","ItemLocation"]]        #Just keep the library locations
location_df = location_df.drop_duplicates()                  #remove duplicates (a single book in multiple 
location_df.head()                                           #locations within the same branch

len(location_df)    #992,518 Records

#### Make Dummy List of all Branch Locations

location_dummies_df = pd.get_dummies(location_df["ItemLocation"])   #branch dummies locations
len(location_dummies_df)                                            #992,518 Records

location_dummies_df.head()

#### Merge DataFrames together to create feature table

location_features_df = pd.merge(location_df,location_dummies_df,left_index=True,right_index=True)

location_features_df = location_features_df.groupby(["BibNum"]).sum()
location_features_df = location_features_df.reset_index()
location_features_df.tail(5)

len(location_features_df)          #397,147 Unique Titles

### Export Data

location_features_df.to_csv("../01_Data/06_Features/Location_Features.csv")
