==> student_submissions/projects/mcnulty/dan_hui/Notebooks/09_Create_Features_Dictionary.ipynb <==
## 09_Create_Features_Dictionary

Author: Daniel Hui

License: MIT

This notebook will take the dictionary dataset and create features from the information heldwithin

import pandas as pd
from collections import defaultdict
pd.set_option('display.max_rows', 125)

### Load Book Inventory Data
We only need to load the book inventory to do some EDA. It is not necessary for creating features. It is useful to see how many books in the collection fall into each category.

inventory_df = pd.read_csv('../01_Data/03_Cleaned/Library_Collection_Inventory_jan_2018_clean.csv',index_col=0)

books_df = inventory_df[["BibNum","Title","ISBN","PublicationYear","Subjects","ItemCollection"]]
books_df = books_df.drop_duplicates()
books_df.head()

There are 567,995 unique records in the book inventory. Each title may have multiple records because different copies of the book are held within different subcollections within the library collection

len(books_df)  

There are 397,147 unique titles in the collection

len(books_df["BibNum"].drop_duplicates()) 

### Load Book Code Data Dictionary
This is the basis of what we will use to create features

dictionary_df = pd.read_csv('../01_Data/02_Truncated/Data_Dictionary_Book_Codes.csv',index_col=0)
dictionary_df.head(5)

There are 212 unique book codes in the dictionary

len(dictionary_df)  

### EDA: How many categories does each book fall into? 
I'm curious how many different subcollections each book may fall into

books_df[["BibNum","ItemCollection"]].groupby(["BibNum"]).count().reset_index().sort_values(by="ItemCollection",ascending=False).head(5)

Book with BibNum 2919133 has 7 different categories associated with it. Curious what the book is below: 

excess_code_df = inventory_df[inventory_df["BibNum"]==2919133][["Title","BibNum","ItemCollection"]].drop_duplicates()
excess_code_df = excess_code_df.rename({"ItemCollection":"Code"},axis=1)
pd.merge(excess_code_df,dictionary_df,on="Code",how="left")

### EDA: How Many books are associated with each book code? 

categories_df = books_df.groupby(["ItemCollection"]).count()
categories_df = categories_df.sort_values(["BibNum"],ascending=False)
categories_df = categories_df.reset_index()[["ItemCollection","BibNum"]]
categories_df = categories_df.rename({"BibNum":"Books","ItemCollection":"Code"},axis=1)
len(categories_df)  # There are 104 book Categories used in the collection

Within the collection, there are only 104 codes used. All their details, including how many books are in each

codes_df = pd.merge(categories_df,dictionary_df,on="Code",how="left")
codes_df.head(10)

### EDA: Examining the Category Subgroup Column

codes_df["Category Subgroup"].unique()   #all the categories

How many biography books? 

codes_df[codes_df["Category Subgroup"] == "Biography"]

codes_df[codes_df["Category Subgroup"] == "Biography"]["Books"].sum()   #18,694 Biographies

How many picture books?

codes_df[codes_df["Category Subgroup"] == "Picture"]

codes_df[codes_df["Category Subgroup"] == "Picture"]["Books"].sum()   #23,771 Picture Books

How many Large Print books?

codes_df[codes_df["Category Subgroup"] == "Large Print"]

codes_df[codes_df["Category Subgroup"] == "Large Print"]["Books"].sum()   #11,974 Large Print Books

How Many Holiday Books? 

codes_df[codes_df["Category Subgroup"] == "Holiday"]

codes_df[codes_df["Category Subgroup"] == "Holiday"]["Books"].sum()   #4,838 Holiday Books

### EDA: Explore the Category Group Column

codes_df["Category Group"].unique()    #3 Category Types

How Many Language Books?

codes_df[codes_df["Category Group"] == "Language"]

codes_df[codes_df["Category Group"] == "Language"]["Books"].sum()   #42,489 Books

How Many Non-Fiction Books?

codes_df[codes_df["Category Group"] == "Nonfiction"]["Books"].sum()  #328,129 Books

How Many Fiction Books?

codes_df[codes_df["Category Group"] == "Fiction"]["Books"].sum()      #175,414 Books

### EDA: Explore the other Columns

codes_df["Format Subgroup"].unique()   #Nothing to see here

codes_df["Format Group"].unique()      #Nothing to see here

codes_df["Description"].unique()        #lot of potential features here

Looking at the list, I will want to pull out categories such as teen, children, comic, african american, and mystery

### Features - Process the Description Column

def search(row, description):
    if description in row:
        return 1
    else: return 0

genre_df = pd.DataFrame()
genre_df["Code"] = codes_df["Code"]
genre_df["Description"] = codes_df["Description"].apply(lambda x: x.lower())         #make lowercase
genre_df["Children"] = genre_df["Description"].apply(search, description="child")
genre_df["Teen"] = genre_df["Description"].apply(search, description="teen")
genre_df["AfAm"] = genre_df["Description"].apply(search, description="afam")
genre_df["Mystery"] = genre_df["Description"].apply(search, description="myster")
genre_df["Comic"] = genre_df["Description"].apply(search, description="comic")       #Comic Books
genre_df["Graphic"] = genre_df["Description"].apply(search, description="graphic")   #sometimes 'Graphic Novel'
genre_df["Comic"] = genre_df["Comic"] + genre_df["Graphic"]                          #Combine
genre_df["Comic"] = genre_df["Comic"].apply(lambda x: min(x, 1))                     #cap max at 1
genre_df = genre_df.drop(["Graphic"],axis=1)

genre_df.tail(10)

### Features - Category Subgroup Dummies

subgroup_df = pd.get_dummies(codes_df["Category Subgroup"])
subgroup_df = subgroup_df.merge(codes_df, left_index=True, right_index=True,how='left')
subgroup_df = subgroup_df[["Code","Biography","Large Print","Picture"]]
subgroup_df.head()

### Features - Category Group Dummies

group_df = pd.get_dummies(codes_df["Category Group"])
group_df = group_df.merge(codes_df, left_index=True, right_index=True,how='left')
group_df = group_df[["Code","Fiction","Language","Nonfiction"]]
group_df["Category"] = codes_df["Category Group"]
group_df.head()

### Merge it all Together

code_feature_df = codes_df.merge(group_df,on="Code",how="left")
code_feature_df = code_feature_df.merge(subgroup_df,on="Code",how="left")
code_feature_df = code_feature_df.merge(genre_df,on="Code",how="left")
code_feature_df = code_feature_df[["Code","Description_x","Category","Fiction","Language","Nonfiction",
                                   "Biography","Large Print","Picture","Children","Teen","Mystery","AfAm",
                                   "Comic"]]
code_feature_df = code_feature_df.rename({"Description_x":"Description"},axis=1)

code_feature_df.head(5)

code_feature_df.to_csv("../01_Data/06_Features/Dictionary_Features.csv")
