==> student_submissions/projects/mcnulty/dan_hui/Notebooks/10_Create_Features_BookTitles.ipynb <==
## 10_Create_Features_BookTitles

Author: Daniel Hui

License: MIT

This notebook creates vectorized word features from the book titles

import pandas as pd
import numpy as np

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer

import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)

pd.set_option('display.max_columns', 500)
pd.set_option('display.max_rows', 500)
pd.set_option('display.float_format', lambda x: '%.4f' % x)

### Import Common Stop Words

import nltk
# nltk.download('stopwords')
from nltk.corpus import stopwords

eng_stopwords = list(stopwords.words('english'))

### Import Book Collection

inventory_df = pd.read_csv(f'../01_Data/03_Cleaned/Library_Collection_Inventory_jan_2018_clean.csv',index_col=0)

book_df = inventory_df[["BibNum","Title"]]
book_df["Title"] = book_df["Title"].fillna("None")                       #drop blank titles
book_df["Title"] = book_df["Title"].apply(lambda x: x.split(" / ")[0])   #this cleans away title cases where 
book_df = book_df.drop_duplicates()
book_df.head()

len(book_df)

### Use TfidfVectorizer to Create Features from Book Titles

Run the TfidfVectorizer function on the book titles

vectorizer = TfidfVectorizer(ngram_range=(1,1),stop_words=eng_stopwords,min_df=.0005,max_df=.016,max_features=150)
X = vectorizer.fit_transform(book_df["Title"])
X.shape

This code is just to peek into what kinds of words are pulled out, ranked from top to bottom

n_grams = list(vectorizer.get_feature_names())
counts = list(sum(X.toarray()))
word_counts = list(zip(n_grams, counts))
pd.DataFrame(word_counts).sort_values(by=1,ascending=False).head(10)

Turn the output array into a dataframe

word_vector_df = pd.DataFrame(X.toarray())
word_vector_df.head(5)

len(word_vector_df)

Rename columns to descriptive columns

word_vector_df.columns = vectorizer.get_feature_names()
word_vector_df.head()

Merge DataFrame back with the unique BibCode

word_vector_df = book_df.merge(word_vector_df,left_index=True, right_index=True, how="left")
word_vector_df.head()

len(word_vector_df)

### Export to CSV

word_vector_df.to_csv("../01_Data/06_Features/Book_Title_Vectors.csv")
