==> student_submissions/projects/mcnulty/dan_hui/Notebooks/17_Clean_Scraped_Data_Random_Sample.ipynb <==
## 17_Clean_Scraped_Data_Random_Sample

Author: Daniel Hui

License: MIT

This notebook cleans the scraped data from the Seattle Public Library's online catalog

import pandas as pd
import numpy as np

# visualization imports
import matplotlib.pyplot as plt
import seaborn as sns
plt.style.use('fivethirtyeight')

import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)

pd.set_option('display.max_columns', 500)
pd.set_option('display.float_format', lambda x: '%.4f' % x)

### Load Scraped Book Data

scrape_df = pd.read_csv('../01_Data/04_Scraped/book_data_random.csv',index_col=0)

scrape_df = scrape_df.reset_index(drop=True)
scrape_df.head()

### Truncate DataSet

scrape_df = scrape_df[["isbn","page","dim","avg_rating","tot_ratings","tot_reviews"]]
scrape_df = scrape_df.rename({"isbn":"ISBN"},axis=1)
scrape_df.head()

scrape_df.info()

### Clean Object Types to Numerical

def make_num(row):
    try:
        return round(row)
    except: -1   #If this happens there was an error

scrape_df["page"] = scrape_df["page"].apply(make_num)
scrape_df["dim"] = scrape_df["dim"].apply(make_num)
scrape_df = scrape_df.fillna(0)

scrape_df.info()

scrape_df[scrape_df["page"] == -1]     #Error Check. Should be Empty

scrape_df[scrape_df["dim"] == -1]     #Error Check. Should be Empty

### Export CSV

scrape_df.to_csv("../01_Data/03_Cleaned/Clean_Book_Data_Random.csv")
