==> student_submissions/projects/mcnulty/dan_hui/Notebooks/19_Model_XGBoost_Scraped_Features.ipynb <==
## 19_Model_XGBoost_Scraped_Features

Author: Daniel Hui

License: MIT

This model uses the core book features, and adds in additional features from the webscraping of the online card catalog to create an XGBoost model

import pandas as pd
import numpy as np

# visualization imports
import matplotlib.pyplot as plt
import seaborn as sns
plt.style.use('fivethirtyeight')

import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)

pd.set_option('display.max_columns', 500)
pd.set_option('display.float_format', lambda x: '%.4f' % x)

### Global Variables
These names correspond to different data sets and targets. Comment them in/out to test different target ranges

target_range = "18_Half"

### Load Main Dataset

checkout_target_df = pd.read_csv(f'../01_Data/07_Merged_Target_Feature_Data/{target_range}_set.csv',index_col=0)

checkout_target_df.head(3)

checkout_target_df.describe()

Keep only the features connected to the book itself, and remove collection and Seattle features

checkout_target_df = checkout_target_df[["BibNum","Checkout","Title","ISBN","Year","Fiction",
                                         "Language","Nonfiction","Biography","Large Print","Picture",
                                         "Children","Teen","Mystery","AfAm","Comic"]]

checkout_target_df.info()

len(checkout_target_df)

### Load Inventory Data Set
I need this for the ISBNs only. And only in the locations I am interested in

isbn_df = pd.read_csv('../01_Data/03_Cleaned/Library_Collection_Inventory_jan_2018_clean.csv',index_col=0)

isbn_df = isbn_df[["BibNum","ISBN"]].drop_duplicates()                 #keep only columns I want
isbn_df.head()

isbn_df.describe()

len(isbn_df)   #We're looking at 30,928 Books

### Load Scraped Book Data

scrape_df = pd.read_csv('../01_Data/03_Cleaned/Clean_Book_Data_Random.csv',index_col=0)

scrape_df.head()

Make total reviews and total ratings integers

scrape_df["tot_ratings"] = scrape_df["tot_ratings"].astype(int)
scrape_df["tot_reviews"] = scrape_df["tot_reviews"].astype(int)
scrape_df = scrape_df.rename({"ISBN":"isbn"},axis=1)
scrape_df.head()

scrape_df.info()

scrape_df.describe()

### Merge 1: Main Set + ISBNS

merge_df = isbn_df.merge(checkout_target_df,on="BibNum",how="left")    #Merge from the ISBN list so it's just
merge_df = merge_df.rename({"ISBN_y":"ISBN","ISBN_x":"isbn"},axis=1)   #those branch locations
merge_df.head()

len(merge_df)   #Should be 30,928

### Merge 2: + Scraped Data

merge_df = merge_df.merge(scrape_df,on="isbn",how="left")
merge_df.head(5)

merge_df.info()

merge_df = merge_df[merge_df["dim"].notnull()]   #remove the data that is not the boosk I
                                                 #scraped

### Train / Test / Split

from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_val_score
from sklearn.neighbors import KNeighborsClassifier
from sklearn.metrics import confusion_matrix
from sklearn import metrics

y = merge_df["Checkout"]
X = merge_df.drop(["Checkout","BibNum","isbn"],axis=1)

#hold out portion of the data for final testing
X, X_test, y, y_test = train_test_split(X, y, test_size=.2, random_state=20) #keep at 20 to be consistent

#hold out 20% for validation
X_train, X_val, y_train, y_val = train_test_split(X, y, test_size=.25, random_state=50) 

print(len(X_train),len(X_val),len(X_test))

X_train.head()

### Logistic Regression

from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score
from sklearn.metrics import recall_score
from sklearn.metrics import precision_score
from sklearn.metrics import f1_score

std_scale = StandardScaler()
X_train_scaled = std_scale.fit_transform(X_train)
lr_model = LogisticRegression(C=10000)

lr_model.fit(X_train_scaled,y_train)

y_train_predict = lr_model.predict(X_train_scaled)
print("Train Accuracy: ", accuracy_score(y_train, y_train_predict ))
print("Train Recall: ",recall_score(y_train, y_train_predict))         # Recall
print("Train Precision: ",precision_score(y_train, y_train_predict))   # Precision
print("Train f1: ",f1_score(y_train, y_train_predict)) 

X_val_scaled = std_scale.transform(X_val)
y_val_pred = lr_model.predict(X_val_scaled)
print("Accuracy: ",metrics.accuracy_score(y_val, y_val_pred))
print("Precision: ",metrics.precision_score(y_val, y_val_pred))
print("Recall: ",metrics.recall_score(y_val, y_val_pred))
print("F1: ",metrics.f1_score(y_val, y_val_pred))

confusion_matrix(y_val, y_val_pred)

Logistic vs XGBoost Conclusion. The Performance of XGBoost was notably better

### XGBoost

import xgboost as xgb

gbm = xgb.XGBClassifier( 
                        n_estimators=30000,
                        max_depth=12,
                        objective='binary:logistic', #new objective
                        learning_rate=.05, 
                        subsample=.8,
                        min_child_weight=3,
                        colsample_bytree=.8
                       )

eval_set=[(X_train,y_train),(X_val,y_val)]
fit_model = gbm.fit( 
                    X_train, y_train, 
                    eval_set=eval_set,
                    eval_metric='error', #new evaluation metric: classification error (could also use AUC, e.g.)
                    early_stopping_rounds=50,
                    verbose=False
                   )

y_train_predict = fit_model.predict(X_train)
print("Train Accuracy: ", accuracy_score(y_train, y_train_predict ))
print("Train Recall: ",recall_score(y_train, y_train_predict))         # Recall
print("Train Precision: ",precision_score(y_train, y_train_predict))   # Precision
print("Train f1: ",f1_score(y_train, y_train_predict)) 

y_val_predict = fit_model.predict(X_val)
print("Validation Accuracy: ", accuracy_score(y_val, y_val_predict ))
print("Validation Recall: ",recall_score(y_val, y_val_predict))         # Recall
print("Validation Precision: ",precision_score(y_val, y_val_predict))   # Precision
print("Validation f1: ",f1_score(y_val, y_val_predict)) 

confusion_matrix(y_val, y_val_predict)
